requirements.txt
streamlit
Pillow
numpy
paddlepaddle
paddleocr
opencv-python-headless